Adding results of manual validation of Reaper
The new validation.csv contains the two authors' responses for each repository manually investigated. The effective classification was reached for each project by deliberation between the two authors, and is used as the ground truth. Then, we show where `reaper` and the stars based classification schemes agree and disagree with the ground truth.